server/*: numerous upsd commands: use size_t where it should be, for numargs parsing

+++ server/netcmds.h
@@ -45,7 +45,7 @@ extern "C" {
 
 static struct {
 	const	char	*name;
-	void	(*func)(nut_ctype_t *client, int numargs, const char **arg);
+	void	(*func)(nut_ctype_t *client, size_t numargs, const char **arg);
 	int	flags;
 } netcmds[] = {
 	{ "VER",	net_ver,	0		},
@@ -68,7 +68,7 @@ static struct {
 	{ "SET",	net_set,	FLAG_USER	},
 	{ "INSTCMD",	net_instcmd,	FLAG_USER	},
 
-	{ NULL,		(void(*)(struct nut_ctype_s *, int,  const char **))(NULL), 0		}
+	{ NULL,		(void(*)(struct nut_ctype_s *, size_t,  const char **))(NULL), 0		}
 };
 
 #ifdef __cplusplus
